Test

See merge request jsql/jsql-api!64

+++ src/main/java/pl/jsql/api/dto/response/QueriesResponse.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public class QueriesResponse {
 
-    public Integer totalQueries;
-    public Integer todayQueries;
+    public Long totalQueries;
+    public Long todayQueries;
 
     public List<QueryResponse> queries;
